+++ test/report/steps.ts
@@ -1,10 +1,11 @@
 import { createBdd } from '../../dist';
 import { test } from './fixtures';
 
-const { Given } = createBdd(test);
+const { Given, When } = createBdd(test);
 
 Given('I am on home page', async ({ myPage }) => {
   await myPage.open();
 });
 
-Given('I do something', async () => {});
+When('Action {int}', () => {});
+When('Состояние {int}', () => {});